refactor(cli): split large Elixir files by concern
Step 1.5 of multi-harness support (sessions#50). Pure reorganization,
no behavior change, tests green.

Code splits:

- cli.ex (327 lines) → cli.ex (151: entry + arg parsing + help)
  + cli/engine.ex (146: port spawn, stream_output loop, finalize_buffer)
  + cli/usage_report.ex (47: run-metrics rendering)

- harness/pi.ex (366 lines) → harness/pi.ex (22: thin doc module)
  + harness/pi/command.ex (78: build_command + default_model)
  + harness/pi/stream.ex (193: process_line, extract_partial_text,
                              flush_partial_buffer, private handlers)
  + harness/pi/tool_input.ex (113: format_tool_input + helpers)

Test splits (mirror code layout):

- cli_test.exs (899 lines) → cli_test.exs (85: arg parsing + module smoke)
  + text_test.exs (72: Cli.Text.check_abort_signal + text_beyond_flushed)
  + harness/pi/stream_test.exs (570: process_line + partial-text)
  + harness/pi/tool_input_test.exs (165: format_tool_input)

Also drops the Cli.process_line / Cli.format_tool_input /
Cli.extract_partial_text / Cli.flush_partial_buffer defdelegate
wrappers — tests now call the sub-module functions directly, so the
compat shims are unnecessary.

Dropped 2 'JSON parsing' tests that exercised Jason.decode/1 directly
(tested the library, not our code).

Why split now: step 5 (claude-code stream parser) will mirror this
structure — Cli.Harness.Claude.{Command,Stream,ToolInput} in parallel.
Symmetric sub-modules make the dispatcher in step 2 obvious and keep
each harness's seams parallel. A monolithic harness/pi.ex would force
the claude adapter to copy the monolith.

Tests: 154/154 bats + 4 doctests + 71 elixir tests, 0 failures.
